refactor(test): move tests to api module
This test organisation allows to reuse code without constant reporting
of unused code
